requirements.txt

azure-core==1.30.2
azure-cosmos==4.7.0
azure-functions==1.20.0
azure-storage-blob==12.14.1
pymongo==4.8.0
